item/pyproject.toml
[project]
name = "item"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "python-jose>=3.3.0",
    "sqlmodel",
]
